move device api functions to auth
